Added general ApplePayEventReceiver to Tester
