+++ Assets/Plugins/iOS/Shopify/BuyTests/ApplePayEventDispatcherTests.swift
@@ -43,6 +43,15 @@ class ApplePayEventDispatcherTests: XCTestCase {
             }
             self.wait(for: [didLoad], timeout: timeout)
         }
+        
+        
+        let setupMessage = UnityMessage(content: "", object: Tester.name, method: Tester.Method.setupApplePayEventTest.rawValue)
+        let messageExpectation = self.expectation(description: "MessageCenter.send(setupApplePayEventTest) failed to complete")
+        MessageCenter.send(setupMessage) { response in
+            messageExpectation.fulfill()
+        }
+        
+        self.wait(for: [messageExpectation], timeout: timeout)
     }
     
     override func tearDown() {
