+++ Assets/UnityTestTools/NativeTesting/Native/Tester.cs
@@ -5,11 +5,19 @@
     using System.Runtime.InteropServices;
 
     public class Tester : MonoBehaviour {
+
         [DllImport ("__Internal")]
         protected static extern void _TesterObjectFinishedLoading();
 
-        void Start () {
-            _TesterObjectFinishedLoading();
+        private bool DidFinishLoading = false;
+
+        void Update() {
+            // We call this in Update() instead of Start() to give a chance
+            // for all tester scripts to initialize in Start()
+            if (!DidFinishLoading) {
+                DidFinishLoading = true;
+                _TesterObjectFinishedLoading();
+            }
         }
     }
 }
