+++ Assets/UnityTestTools/NativeTesting/iOS/ApplePayEventTesterReceiver.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 20537e7048a7548edbcbc7ff5df33fdf
+timeCreated: 1497623880
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
